fix: Updated "Remember Me" functionality
Added functionality to remember logged-in users using cookies.

Introduced a new filter to check for the "rememberMe" cookie and automatically log in the user if a valid token is found.

Created a SessionAuthToken entity to store authentication tokens, along with corresponding repository and service classes.

Modified the sign-in servlet to create and set the "rememberMe" cookie upon successful login.

Updated the sign-out servlet to delete the "rememberMe" cookie and invalidate the session.

Removed the WebFilter annotation from AuthenticationFilter and registered both filters in web.xml.